Update Main_Menu.asm
Updated The Nomenclature In Main_Menu:, Calling The Loop Instead Of The Start/Initialization Of The File. This Fixed The Issue With Cursor_Flags And Character_Flags/Stage_Flags Resetting To Zero Quickly; Note To Add A Blurb About This In The Article, Since It Had `Callf Main_Menu` In `Main.ASM` At First On Purpose.