Update README.md
Added A New Header For "Coding The Menu Variables And Inputs," To Be Used In Conjunction With "Drawing The Menu Text" And To Break Up The Massive Text Block Forming That Section, Helping With Readability For Users. Will Want To Consider Moving Text Around And Possibly Introducing "Check_Button_Pressed" Later If Starting With The "Graphics Drawing" Part Is Preferable.